fix delete activity verify

+++ src/controller.php
@@ -25,40 +25,46 @@ function controller() {
                     case 'inbox':
                         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                             $jsonld = json_decode($input = file_get_contents('php://input'), 1);
-                            if (isset($jsonld['actor']) && parse_url($jsonld['actor'])['host'] != $config['base'] &&
-                            ($jsonld['type'] == 'Delete' || $actor = Club_Get_Actor($club, $jsonld['actor']))) {
+                            if (isset($jsonld['actor']) && parse_url($jsonld['actor'])['host'] != $config['base']) {
+                                
+                                $verify = ActivityPub_Verification($input);
+                                if ($jsonld['type'] == 'Delete' && $jsonld['actor'] == $jsonld['object']) {
+                                    if ($verify) {
+                                        $pdo = $db->prepare('delete from `users` where `actor` = :actor');
+                                        $pdo->execute([':actor' => $jsonld['actor']]);
+                                    } break;
+                                }
                                 if ($config['nodeDebugging']) {
                                     $file_name = date('Y-m-d_H:i:s_').$club.'_'.$jsonld['type'];
-                                    if (!($jsonld['type'] == 'Delete' && $jsonld['actor'] == $jsonld['object'])) {
-                                        file_put_contents('inbox_logs/'.$file_name.'_input.json', $input);
-                                        file_put_contents('inbox_logs/'.$file_name.'_server.json', Club_Json_Encode($_SERVER));
-                                    }
+                                    file_put_contents('inbox_logs/'.$file_name.'_input.json', $input);
+                                    file_put_contents('inbox_logs/'.$file_name.'_server.json', Club_Json_Encode($_SERVER));
+                                    if (!$verify) file_put_contents('inbox_logs/'.$file_name.'_verify_failed', '');
                                 }
-                                if (!ActivityPub_Verification($input)) {
-                                    if ($config['nodeDebugging']) file_put_contents('inbox_logs/'.$file_name.'_verify_failed', '');
-                                    break;
-                                }
+                                if (!$verify) break;
+                                
                                 switch ($jsonld['type']) {
                                     case 'Create': Club_Announce_Process($jsonld); break;
                                     case 'Follow':
-                                        $pdo = $db->prepare('insert into `followers`(`cid`,`uid`,`timestamp`) select `cid`, :uid as `uid`, :timestamp as `timestamp` from `clubs` where `name` = :club');
-                                        $pdo->execute([':club' => $club, ':uid' => $actor['uid'], ':timestamp' => time()]);
-                                        $pdo = $db->prepare('select f.id from `followers` as f left join `clubs` as `c` on f.cid = c.cid where f.uid = :uid and c.name = :club');
-                                        $pdo->execute([':club' => $club, ':uid' => $actor['uid']]);
-                                        if ($follow_id = $pdo->fetch(PDO::FETCH_COLUMN, 0)) {
-                                            Club_Push_Activity($club, [
-                                                '@context' => 'https://www.w3.org/ns/activitystreams',
-                                                'id' => $club_url.'#accepts/follows/'.$follow_id,
-                                                'type' => 'Accept',
-                                                'actor' => $club_url,
-                                                'object' => [
-                                                    'id' => $jsonld['id'],
-                                                    'type' => 'Follow',
-                                                    'actor' => $jsonld['actor'],
-                                                    'object' => $club_url
-                                                ]
-                                            ], $actor['inbox']);
-                                        } break;
+                                        if ($actor = Club_Get_Actor($club, $jsonld['actor'])) {
+                                            $pdo = $db->prepare('insert into `followers`(`cid`,`uid`,`timestamp`) select `cid`, :uid as `uid`, :timestamp as `timestamp` from `clubs` where `name` = :club');
+                                            $pdo->execute([':club' => $club, ':uid' => $actor['uid'], ':timestamp' => time()]);
+                                            $pdo = $db->prepare('select f.id from `followers` as f left join `clubs` as `c` on f.cid = c.cid where f.uid = :uid and c.name = :club');
+                                            $pdo->execute([':club' => $club, ':uid' => $actor['uid']]);
+                                            if ($follow_id = $pdo->fetch(PDO::FETCH_COLUMN, 0)) {
+                                                Club_Push_Activity($club, [
+                                                    '@context' => 'https://www.w3.org/ns/activitystreams',
+                                                    'id' => $club_url.'#accepts/follows/'.$follow_id,
+                                                    'type' => 'Accept',
+                                                    'actor' => $club_url,
+                                                    'object' => [
+                                                        'id' => $jsonld['id'],
+                                                        'type' => 'Follow',
+                                                        'actor' => $jsonld['actor'],
+                                                        'object' => $club_url
+                                                    ]
+                                                ], $actor['inbox']);
+                                            }
+                                        } else Club_Json_Output(['message' => 'Actor not found'], 0, 400); break;
                                     case 'Undo':
                                         switch ($jsonld['object']['type']) {
                                             case 'Follow':
@@ -67,16 +73,12 @@ function controller() {
                                                     ' (select cid from `clubs` where `name` = :club)'.
                                                     ' and `uid` in (select uid from `users` where `actor` = :actor)');
                                                 $pdo->execute([':club' => $club, ':actor' => $jsonld['actor']]); break;
-                                                
                                             default: break;
                                         } break;
                                     case 'Delete':
                                         if (isset($jsonld['object']['type'])) switch ($jsonld['object']['type']) {
                                             case 'Tombstone': Club_Tombstone_Process($jsonld); break;
                                             default: break;
-                                        } elseif ($jsonld['actor'] == $jsonld['object']) {
-                                            $pdo = $db->prepare('delete from `users` where `actor` = :actor');
-                                            $pdo->execute([':actor' => $jsonld['actor']]);
                                         } else {
                                             $jsonld['object'] = ['id' => $jsonld['object']];
                                             Club_Tombstone_Process($jsonld);
@@ -258,26 +260,28 @@ function controller() {
             if ($_SERVER['REQUEST_METHOD'] == 'POST') {
                 $jsonld = json_decode($input = file_get_contents('php://input'), 1);
                 if (isset($jsonld['actor']) && parse_url($jsonld['actor'])['host'] != $config['base']) {
+                    
+                    $verify = ActivityPub_Verification($input);
+                    if ($jsonld['type'] == 'Delete' && $jsonld['actor'] == $jsonld['object']) {
+                        if ($verify) {
+                            $pdo = $db->prepare('delete from `users` where `actor` = :actor');
+                            $pdo->execute([':actor' => $jsonld['actor']]);
+                        } break;
+                    }
                     if ($config['nodeDebugging']) {
                         $file_name = date('Y-m-d_H:i:s').'_shared_inbox_'.$jsonld['type'];
-                        if (!($jsonld['type'] == 'Delete' && $jsonld['actor'] == $jsonld['object'])) {
-                            file_put_contents('inbox_logs/'.$file_name.'_input.json', $input);
-                            file_put_contents('inbox_logs/'.$file_name.'_server.json', Club_Json_Encode($_SERVER));
-                        }
+                        file_put_contents('inbox_logs/'.$file_name.'_input.json', $input);
+                        file_put_contents('inbox_logs/'.$file_name.'_server.json', Club_Json_Encode($_SERVER));
+                        if (!$verify) file_put_contents('inbox_logs/'.$file_name.'_verify_failed', '');
                     }
-                    if (!ActivityPub_Verification($input)) {
-                        if ($config['nodeDebugging']) file_put_contents('inbox_logs/'.$file_name.'_verify_failed', '');
-                        break;
-                    }
+                    if (!$verify) break;
+                    
                     switch ($jsonld['type']) {
                         case 'Create': Club_Announce_Process($jsonld); break;
                         case 'Delete':
                             if (isset($jsonld['object']['type'])) switch ($jsonld['object']['type']) {
                                 case 'Tombstone': Club_Tombstone_Process($jsonld); break;
                                 default: break;
-                            } elseif ($jsonld['actor'] == $jsonld['object']) {
-                                $pdo = $db->prepare('delete from `users` where `actor` = :actor');
-                                $pdo->execute([':actor' => $jsonld['actor']]);
                             } else {
                                 $jsonld['object'] = ['id' => $jsonld['object']];
                                 Club_Tombstone_Process($jsonld);
